qtwebengine: add workaround for x86-64 targets
Some x86-64 target, such as intel-skylake-64, use AVX instructions that are
not supported by QEMU (https://gitlab.com/qemu-project/qemu/-/issues/164).
This causes the build tools to crash with QEMU when they are ran during
to build. Instead of using QEMU, change the wrapper to try to run the
tools directly, using the target dynloader.

Change-Id: I790e013f91ce5f0e42fb1e11ff61e4428d14d9ad
Reviewed-by: Pasi Petäjäjärvi <pasi.petajajarvi@qt.io>