Coin: build Qt packagegroups in CI integrations
Current CI build using the toolchain recipe is a bit limited as the
package_qa tasks were never ran for the Qt recipes. Also the rm_work
was not done which lead to disk space running out in CI builds. Add
both native and target packagegroup recipes to the target builds that
will ensure that both of the above mentioned tasks are now run.

Pick-to: 6.2
Change-Id: Ifb73b807f4f9f551836bfa20003aaf88bc758d78
Reviewed-by: Pasi Petäjäjärvi <pasi.petajajarvi@qt.io>