qtmultimedia: update PACKAGECONFIGs
QtMultimedia is again using plugins for different backends
and they can be enabled at the same time. Add PACKAGECONFIGs
for ffmpeg and vaapi backends. Remove vivante workaround,
it's no longer needed.

Default still to gstreamer only. The ffmpeg integration is done
to v4.4 and does not work with later versions.

Change-Id: Ica8f46ceca0ef61caa300b3e23269de062575efc
Reviewed-by: Mikko Gronoff <mikko.gronoff@qt.io>