layer: use the same priority as the core layer
Some of the python3 recipes we carry are already in meta-python
layer, but not in all the branches that we support. In order to use
the most recent recipe, we need to have same or lower BBFILE_PRIORITY
than meta-python. Depending on the branch, meta-python's priority is
either same or higher than the core layer's priority.

Change-Id: I98c74858c5be69e45a76a45ed10aa1168c614618
Reviewed-by: Mikko Gronoff <mikko.gronoff@qt.io>
(cherry picked from commit d36033ca56e96d8843d9e733b50759f237021694)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>