Fix minizip usage in qtwebengine
WebEngine needs minizip which is part of zlib.Till now it always built
bundled zip as minizip recipe does not exist in poky and moreover this
was never a configurable feature as it was not in PACKAGECONFIG.

A change in https://codereview.qt-project.org/c/qt/qtwebengine/+/384262
enables compilation of bundled minizip against system zlib (by using
shim headers) however it is not advisable as those two version of zip
might be different and minizip is project wise a part of zlib.

Therefore introduce cheap rip off of zlib recipe, however
building minizip from same source tree.

Task-number: QTBUG-103149
Pick-to: 6.4 6.3 6.2
Change-Id: I1e36d2fba339fb5383018f05a6347c27dde0c8c1
Reviewed-by: Samuli Piippo <samuli.piippo@qt.io>