sdk: move sdk mkspec additions to qtbase recipe
The qmake mkspec additions are used with the sdk when the toolchain
environment-setup script hasn't been run. Instead of adding those in
the sdk creation phase, do it already in the qtbase recipe.

Change the mkspec also so that CC/CXX has only the compiler executable,
but none of the compile flags that usually are there in Yocto builds.
This keeps Qt Creator happier as well.

Pick-to: 6.5 6.4 6.2
Task-number: QTCREATORBUG-27372
Change-Id: If27d595965df59ff9846f4c773469889a8480040
Reviewed-by: Mikko Gronoff <mikko.gronoff@qt.io>
Reviewed-by: Ari Parkkila <ari.parkkila@qt.io>